requirements.txt
fastapi
uvicorn
sqlalchemy
feedparser
openai
psycopg2-binary
pydantic>=2.0
python-dotenv
python-dateutil